Rename enumerateNode util
